# docs/swagger.yaml
definitions:
...
  mengri-flow_internal_app_dto.CreateResourceRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
      type:
        type: string
      workspaceId:
        type: string
    required:
    - config
    - name
    - type
    - workspaceId
    type: object
...
  mengri-flow_internal_app_dto.ExecutionTimelineResponse:
    properties:
      completedAt:
        type: string
      error:
        type: string
      executionTime:
        type: integer
      nodeId:
        type: string
      nodeName:
        type: string
      startedAt:
        type: string
      status:
        type: string
      toolId:
        type: string
      toolName:
        type: string
    type: object
...
  mengri-flow_internal_app_dto.ListResourcesResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.ResourceResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  mengri-flow_internal_app_dto.ListRunsResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.RunResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  mengri-flow_internal_app_dto.ResourceResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  mengri-flow_internal_app_dto.RunDetailResponse:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      flowId:
        type: string
      flowName:
        type: string
      id:
        type: string
      input:
        additionalProperties: true
        type: object
      nodeCount:
        type: integer
      nodeStatus:
        additionalProperties:
          type: string
        type: object
      output:
        additionalProperties: true
        type: object
      startedAt:
        type: string
      status:
        type: string
      triggerId:
        type: string
      triggerType:
        type: string
    type: object
  mengri-flow_internal_app_dto.RunResponse:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      flowId:
        type: string
      flowName:
        type: string
      id:
        type: string
      startedAt:
        type: string
      status:
        type: string
      triggerId:
        type: string
      triggerType:
        type: string
    type: object
  mengri-flow_internal_app_dto.RunStatsResponse:
    properties:
      failedRuns:
        type: integer
      monthRuns:
        type: integer
      runningRuns:
        type: integer
      successRuns:
        type: integer
      todayRuns:
        type: integer
      totalRuns:
        type: integer
      weekRuns:
        type: integer
    type: object
...
  mengri-flow_internal_app_dto.TestConnectionRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      type:
        type: string
    required:
    - config
    - type
    type: object
...
  mengri-flow_internal_app_dto.UpdateResourceRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
    type: object
...
  mengri-flow_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /resources:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: query
        name: workspaceId
        required: true
        type: string
      - description: 资源类型
        in: query
        name: type
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListResourcesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取资源列表
      tags:
      - Resource
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建资源请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.CreateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ResourceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建资源
      tags:
      - Resource
  /resources/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 资源ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除资源
      tags:
      - Resource
    get:
      consumes:
      - application/json
      parameters:
      - description: 资源ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ResourceResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取资源详情
      tags:
      - Resource
    put:
      consumes:
      - application/json
      parameters:
      - description: 资源ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新资源请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.UpdateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ResourceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新资源
      tags:
      - Resource
  /resources/test-connection:
    post:
      consumes:
      - application/json
      parameters:
      - description: 测试连接请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.TestConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 测试资源连接
      tags:
      - Resource
  /runs:
    get:
      consumes:
      - application/json
      parameters:
      - description: 流程ID
        in: query
        name: flowId
        type: string
      - description: 触发器ID
        in: query
        name: triggerId
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListRunsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取运行记录列表
      tags:
      - Run
  /runs/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 运行ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.RunDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取运行详情
      tags:
      - Run
  /runs/{id}/retry:
    post:
      consumes:
      - application/json
      parameters:
      - description: 运行ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.RunResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 重试运行
      tags:
      - Run
  /runs/{id}/timeline:
    get:
      consumes:
      - application/json
      parameters:
      - description: 运行ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ExecutionTimelineResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取执行时间线
      tags:
      - Run
  /runs/stats:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.RunStatsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取运行统计
      tags:
      - Run
...
swagger: "2.0"